+++ migrations/20260529091751_join_leave.sql
@@ -1,7 +1,8 @@
-CREATE TABLE join_leave_logs (
-    id SERIAL PRIMARY KEY,
-    user_id BIGINT NOT NULL,
-    guild_id BIGINT NOT NULL,
-    action VARCHAR(10) NOT NULL, -- 'JOIN' or 'LEAVE'
+CREATE TABLE join_leave_logs
+(
+    id         BIGSERIAL PRIMARY KEY,
+    user_id    BIGINT      NOT NULL,
+    guild_id   BIGINT      NOT NULL,
+    action     VARCHAR(10) NOT NULL, -- 'JOIN' or 'LEAVE'
     created_at TIMESTAMPTZ DEFAULT NOW()
 );
